knowledge_manager/requirements.txt
fastapi==0.115.14
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.59
langchain-milvus==0.1.9
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
markitdown==0.1.1
pymupdf4llm==0.0.24
python-dotenv==1.1.0
uvicorn==0.35.0